refactor(doctor): extract types.rs, add default Check::fix
